backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
codecarbon==2.3.4
python-multipart==0.0.9
python-dotenv==1.0.1
